Reordered gap/border placement. Abstracted gap reg
